% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCropArea.R
\name{addCropArea}
\alias{addCropArea}
\title{addCropArea}
\usage{
addCropArea(pathIn, shapeAdd, cropAmt, pathOut = "")
}
\arguments{
\item{pathIn}{The path of the parsed and projected crop nc file}

\item{shapeAdd}{Either A shape file from Google Earth (kml) defining the area,
or an x,y matrix defining the added points.}

\item{cropAmt}{The amount of crop in that area which is either constant or 
the length of the input shape file}

\item{pathOut}{The path out of the final nc file}
}
\value{
The final raster and the new nc file if pathOut is specified.
}
\description{
This function is for adding an additional crop area that was not included in 
the NASS system. The most obvious use case is for areas just outside of U.S, 
like in the Ontario province in Canada. Another is for transcontinental
migration adding large splotches of Mexico, Carribean and Canada. The 
function works by looking at a shape file for a grid of points (lat, lon) and
using the cropAmt to create an additional area to add to the original file.
}

