% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprioriVars.R
\name{aprioriVars}
\alias{aprioriVars}
\title{Turn the data into abstracted variables for the model.}
\usage{
aprioriVars(dirTrees, pathOut, plantTimes, harvestTimes, CornThres,
  TGDDbaseCrop, TGDDbaseFAW, windTOThres, tempTOThres, precTOThres)
}
\arguments{
\item{dirTrees}{a list of all the character vectors of where the data lives}

\item{pathOut}{the path where the ncdf file will be located}

\item{plantTimes}{a vector of length 4, that defines the three sections of
corn development. The first two numbers are latitudes that divide the 
country into three parts. The second two numbers are the planting dates of 
at those latitudes with the dates being interpolated between the two points 
and held constant anywhere else.}

\item{harvestTimes}{a vector with the same stucture as panting dates but for
the time of harvest.}

\item{CornThres}{a hectare amount of recognizable corn in one grid cell.
Represents a minimum desnsity of hectares /1600 km^2}

\item{TGDDbaseCrop}{The base temperature value for the degree day
calculations of corn growth in units of celcius.}

\item{TGDDbaseFAW}{The base temperature value for the degree day
calculations of insect growth in units of celcius.}

\item{windTOThres}{If the wind speed (m/s) is greater than this value then don't fly}

\item{tempTOThres}{If the temp (degC) is lower than this value then don't fly}

\item{precTOThres}{If the precipitation (3hr acc. m) is greater than this value then don't fly}
}
\value{
Works as a byproduct of the saving of the ncdf file.
}
\description{
For computer time, memory and simplicity it is easier to deal with 
the aggregated variables in the actual biological model than the actual data.
The main variable calculated is the growth of corn over the year.
Other variables include Fall Army worm growth and wind mangitude northward.
Other variables that can be included is the decision to fly in the form of
a set of logicals including precipitation, biotic factors or temperature.
The parameters are saved with the variables as global attributes in the 
final ncdf file. This file is loaded at the begining of iterateHYSPLIT
}
\details{
The units for the ouput values are set up to reduce the size of the 
	data. The degree day values are done in 0.1*C*day or 0.1*F*day so
 they can be stored as integers.
}

