% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosplit.R
\docType{package}
\name{biosplit}
\alias{biosplit}
\alias{biosplit-package}
\title{biosplit: A model for national insect migration}
\description{
biosplit package provides functions for modeling the biological growth of 
insects and their food source as well as nightly flight using dispersion 
modeling from the HYSPLIT model from NOAA. Included are all the tools needed
to collect the data required, run the model and analye the results. For more 
information on how all the functions work you can go to vignette/workflow.rmd
}

