% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{changeConfig}
\alias{changeConfig}
\title{Change the configuration file}
\usage{
changeConfig(path = "", ...)
}
\arguments{
\item{path}{The path to the config file if different than the package directory}

\item{...odd}{config element name as a string e.g. "runName", "year".}

\item{...even}{the value specified. e.g. "runWorld", "2012".}
}
\description{
After the configuration path, there should be pairs of configuration
elements that should be changed. The function loads up the file as lines and changes
those values then rewrites the entire file.
}
\examples{
cfg <- loadConfig()
changeConfig(,'year',2014)
changeConfig(,'year',cfg$year,'runName',cfg$runName)

\dontrun{
changeConfig(,'asdjkf',5)
#Error asdjkf are not valid variables in Config
}
}

