% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{loadConfig}
\alias{loadConfig}
\title{Load configuration parameters for the run, biological and simulation needs}
\usage{
loadConfig(path = "")
}
\arguments{
\item{path}{The path to the config file if different than the package
directory}
}
\value{
a list of all the elements in the configuration file with their values
Also sets up the directory trees for the data
Saves the R object as a RData file and a second text file "cfg.txt" 
for other languages
}
\description{
Parses the config file and returns the cfg with all the parameters
as an object, a .RData file and compiled text documents
Also creates all the directories as specified in the config files
}

