% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmapBackups.R
\name{makeDiagnosticMap}
\alias{makeDiagnosticMap}
\title{Make a diagonstic map for a layer}
\usage{
makeDiagnosticMap(arr, type = c("contour", "post")[1], pathSave = "")
}
\arguments{
\item{arr}{Anything that can be loaded in by 'raster' package}

\item{type}{The type of map that you want printe}

\item{pathSave}{The path that you want the saved jpeg to be in. Default is that the function does not save but just plot}
}
\value{
A plot with 2 layers, the base map and the array
}
\description{
A one layer map with the designated layer.
}

