% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScript.R
\name{makeRunFun}
\alias{makeRunFun}
\title{A functionfactory for running scripts in this package}
\usage{
makeRunFun(driver, ext)
}
\arguments{
\item{driver}{The file location of the root folder of the desired function.}

\item{ext}{The extension that the function is to be used with (py, BAS or R)}
}
\description{
An OS agnostic way of calling a script from the command line given the
location of the program file. The scripts are either in the scripts folder or
 the individual languges. The py extension refers to the Jython scripts that 
 are used with MeteoInfo which needs to eb installed beforehand. The BAS 
 extension is used with Golden Software's Surfer's automation tool, scripter.
}

