% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NASS2TIFs.R
\name{NASS2TIFs}
\alias{NASS2TIFs}
\title{Download the Cropscape files}
\usage{
NASS2TIFs(outPath, year, cropExtent, cropType = "Corn")
}
\arguments{
\item{outPath}{The path where the .tifs are saved}

\item{year}{The year to grab}

\item{cropExtent}{An extent of the downloaded area}

\item{cropType}{The crop type you want to export}
}
\value{
Horizontal slices of the area of interest in the path you specify. In R, nothing.
}
\description{
Uses the Cropscape api to get download the crop .tiffs of the selected crop type
}

