% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncdf2trapdata.R
\name{ncdf2trapdata}
\alias{ncdf2trapdata}
\title{Post-processing: Extract model output at trap locations for comparison}
\usage{
ncdf2trapdata(dirSim, pathTrap, useCombined = TRUE, shWrite = TRUE)
}
\arguments{
\item{dirSim}{The simulation output folder}

\item{pathTrap}{The path to the trap location csv file with the
lat/lon of all trap locations}

\item{useCombined}{Should the program use the combined Final.nc file
or the nc slices in the nc folder}

\item{shWrite}{Should the function wrtie the horizontal and
    vertical csv files or just return the horizontal matrix}
}
\value{
shWrite = 1: only writes the csv file
	 shWrite = 0: A matrix of the horizontal time series
	 with the identifying information in front
}
\description{
Extracts the model output by location and outputs the
	time series for the entire year for that location
}

