% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncdf2trapdata.R
\name{ncdf2trapgrid}
\alias{ncdf2trapgrid}
\title{Post-processing: Compare the output with the gridded trap data}
\usage{
ncdf2trapgrid(dirSim, pathTrapGrid, pathOut,
  goodProj = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")
}
\arguments{
\item{dirSim}{The simulation output folder}

\item{pathTrapGrid}{The path to the trap grd location}

\item{pathOut}{The path where the nc file should be saved}

\item{goodProj}{The projection of the target raster}
}
\value{
A raster of the difference between the simulation and the trap data
which is also saved in the path specified as an nc file
}
\description{
Compares the grided first occurance of the traps with the output first occurance
}

