% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rawCrop2nicenc.R
\name{rawCrop2nicenc}
\alias{rawCrop2nicenc}
\title{Manipulate the input slices into a nice netCDF file}
\usage{
rawCrop2nicenc(dirIn, pathOut, cropGrid, niceGrid = "")
}
\arguments{
\item{dirIn}{The directory with slices of tif files ussually generated by NASS2TIFs()}

\item{pathOut}{The full file name of the aggregated file, either in .dat or .nc}

\item{cropGrid}{The input extent and projection with the new spacing that you want}

\item{niceGrid}{The result grid in the projection that you want}
}
\value{
saves the result as a csv readable by surfer (.dat) and the netCDF file at pathOut
	as specified. Returns a 0 if the writing failed, 1 if it worked.
}
\description{
Scales the color tag to a single true or false value,
	aggregates the small spacing into the desired spacing using sum(extract())
	since raster::aggregate can't take uneven spaces.
	Then it changes the projection into unprojected lat/lon or a user defined projection
}

