% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rawMet2nicenc.R
\name{rawMet2nicenc}
\alias{rawMet2nicenc}
\title{Process the daily grid files into a nice netCDF file}
\usage{
rawMet2nicenc(dirTreeIn, projKey, niceGrid, unit = "unitless",
  transCol = c(1, 2))
}
\arguments{
\item{dirTreeIn}{A character array of the three directories to use
    1) where the input files are located
    2) where the projected .nc slices are going to be outputted
    3) where the final combined output .nc file is outputted}

\item{projKey}{A string with the projection info either in the form of
    a file location of a mapping table or proj4string for the input grids}

\item{niceGrid}{A grid of the final extent and projection that is needed}

\item{unit}{A string designating the unit of the variable...use deg for degree}

\item{transCol}{A vector with column numbers for the mapping table, (lat,long),
Only used if projKey is a mapping table file location}
}
\value{
Saves the projected and cropped slices in the second entry in DirTreeIn,
	and two nc files with the stacked nc files for the Highs and the lows for the whole year
}
\description{
Goes through the twice daily grid files for one variable and projects, crops and combines them
	into two managable nc files for the highs and the lows.
	The grids are produced by ARL2GRD.py and the result is used by aprioriVars.R.
}

