% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncdf2trapdata.R
\name{rebuildNc}
\alias{rebuildNc}
\title{Rebuild the sparse matrices with raw data}
\usage{
rebuildNc(dirSim, yr, flagSum = FALSE, shWrite = FALSE)
}
\arguments{
\item{dirSim}{The simulation output folder}

\item{yr}{The year of the simulation}

\item{flagSum}{Should the program sum all the captures for that week in this
file?}

\item{shWrite}{Should the rebuilt file be written to the directory using
  predefined names?}
}
\value{
A large list with the variables as names and 3-d matrix for the data
}
\description{
Rebuild the sparse matrices with raw data
}

