% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iterateHYSPLIT.R
\name{runBiosplit}
\alias{runBiosplit}
\title{Run the biosplit model for a year}
\usage{
runBiosplit(cfg)
}
\arguments{
\item{config}{The configuration object returned by loadConfig
or from loading the cfg.RData in the config.txt location}
}
\value{
Time stamp. if called for in cfg, also writes the weekly snapshots of
	 populations, in .txt and .nc format.
	 It also outputs a final nc file after everything is completed.
}
\description{
Simulates continental moth migration over the course of a year.
	Requires HYSPLIT to be installed.
}

