% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubValidation.R
\name{scrubHaplo}
\alias{scrubHaplo}
\title{Make the haplotype data easier to look at}
\usage{
scrubHaplo(pathXlsx, pathTrapDict, pathCsvOut = "", maskBefore2011 = TRUE,
  onlyYr = c(2011, 2012, 2013, 2014), maskNumMoth = 15)
}
\arguments{
\item{pathXlsx}{The path to the xlsx with the Haplotype data}

\item{pathTrapDict}{A csv with the translation of all the County names to the latLon location}

\item{pathCsvOut}{The csv location where the nice data should be outputed}

\item{maskBefore2011}{should values before 2011 be used?}

\item{onlyYr}{an array of the years you want to be kept}

\item{maskNumMoth}{The minimum number of moths that are required for a proper Haplotype experiment}
}
\value{
The final matrix and a csv file that was written if pathCsvOut was specified
}
\description{
Make the haplotype data easier to look at
}

