% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addCropArea.R
\name{simAreaStats}
\alias{simAreaStats}
\title{simAreaStats: Interrogate the final simulation on specific areas}
\usage{
simAreaStats(dirSim, pathShape, useSum = FALSE)
}
\arguments{
\item{dirSim}{The simulation output folder}

\item{pathShape}{The path to a kml polygon to define the extent}

\item{useSum}{Either uses the final combined sum nc file or creates it using
  "rebuildNc"}
}
\value{
a list of the first occurance for both the populations and the
mixing area for the entire area and the entire year
}
\description{
This post processing process allows the user to analyze regions instead of
the entire US. A number of statistics are outputted as elements in a list.
}

