% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCropArea.R
\name{valAreaStats}
\alias{valAreaStats}
\title{Interrogate the validation data on specific areas}
\usage{
valAreaStats(csvIn, pathShape, niceGrid)
}
\arguments{
\item{csvIn}{Either the path or the matrix from the summarized validation data,
or the output of "summarizeValid()"}

\item{pathShape}{The path to a kml polygon to define the extent}

\item{niceGrid}{a raster or a raster extent of the final entire grid 
that the data is on}
}
\value{
a list of the summary stats
}
\description{
Interrogate the validation data on specific areas
}

