% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{changeConfig}
\alias{changeConfig}
\title{Change the configuration file}
\usage{
changeConfig(..., path = "")
}
\arguments{
\item{...}{odd members are config element name as a string e.g. "runName",
"year". while even values are thevalue to change it to e.g. "runWorld", "2012".}

\item{path}{The path to the config file if different than the package directory}
}
\description{
Specify a option - value pair to be changed in the configuration file.
The function loads up the file as lines and changes
those values then rewrites the entire file.
}
\examples{
cfg <- loadConfig()
changeConfig('year',2014)
changeConfig('year',cfg$year,'runName',cfg$runName)

\dontrun{
changeConfig('asdjkf',5)
#Error asdjkf are not valid variables in Config
}
}

