% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{loadConfig}
\alias{loadConfig}
\title{Load configuration parameters for the run, biological and simulation needs}
\usage{
loadConfig(path = "config.txt")
}
\arguments{
\item{path}{The path to the config file if different than the package directory}
}
\description{
Parses the config file and returns the cfg with all the parameters
as an object, a .RData file and compiled text documents
}

