% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NASS2TIFs.R
\name{NASS2TIFs}
\alias{NASS2TIFs}
\title{Download the Cropscape files}
\usage{
NASS2TIFs(dirOut, year, cropGrid, cropType = "Corn")
}
\arguments{
\item{dirOut}{The directory where the .tifs are saved}

\item{year}{The year to grab from CropScape}

\item{cropGrid}{An grid of the downloaded area, need extent and spacing}

\item{cropType}{The crop type you want to export}
}
\value{
Horizontal slices of the area of interest in the path you specify. In R, nothing.
}
\description{
Uses the Cropscape api to get download the crop .tiffs of the selected crop type
}

