% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf2trapdata.R
\name{ncdf2trapdata}
\alias{ncdf2trapdata}
\title{Post-processing: Extract model output at trap locations for comparison}
\usage{
ncdf2trapdata(dirSim, pathTrap, pathHap = NULL, useCombined = TRUE,
  shUseSum = FALSE, shWrite = TRUE, shAvgSimforHap = FALSE, notes = "")
}
\arguments{
\item{dirSim}{The simulation output folder}

\item{pathTrap}{The path to the trap location csv file with the 
lat/lon of all trap locations or the combined records list with all the data}

\item{pathHap}{The path to the haplotype data in a nice csv output of scrubHap}

\item{useCombined}{Should the program use the combined Final.nc file 
or the nc slices in the nc folder to use do the snap shot calculation}

\item{shUseSum}{Go through the individual slices and sum up for the week 
instead of using the combined file}

\item{shWrite}{Should the function write the horizontal and 
vertical csv files or just return the horizontal matrix}

\item{shAvgSimforHap}{Should the program add an extra column with the average
mixing ratio for those dates where the haplotype record is spread out 
over more than one week?}

\item{notes}{A vector or string of any notes that should be included in the
appendix}
}
\value{
shWrite = 1: only writes the csv file
      shWrite = 0: A matrix of the horizontal time series 
 				with the identifying information in front
}
\description{
Extracts the model output by location and outputs the 
time series for the entire year for that location
}
\details{
For the summation analysis type, sometimes there may be less than 7
	files in the ncs folder due to model conditions (cfg$outEveryDayStart) so 
output whether the week has all the files as a logical 52 week vector along 
with a combined nc file which is used like the regular ncfile throughout.
outputs the summation of less than 7 files anyway.
}

