% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncdf2trapdata.R
\name{ncdf2trapgrid}
\alias{ncdf2trapgrid}
\title{Post-processing: Compare the output with the gridded trap data}
\usage{
ncdf2trapgrid(dirSim, pathTrapGrid, pathOut, goodRas)
}
\arguments{
\item{dirSim}{The simulation output folder}

\item{pathTrapGrid}{The path to the trap grd location}

\item{pathOut}{The path where the nc file should be saved}

\item{goodRas}{A raster with the target grid, projection and extent}
}
\value{
Works as a byproduct by writing a grd of the difference and the xydiff
}
\description{
Compares the grided first occurance of the traps with the output first occurance
}

