% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncdf2trapdata.R
\name{ncdf2trapgrid}
\alias{ncdf2trapgrid}
\title{Post-processing: Summarize the simulation details in a nicely mapped format}
\usage{
ncdf2trapgrid(dirSim, pathTrapGrid = "", pathOut = "",
  goodProj = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
  shDoMix = FALSE, shUseSum = FALSE)
}
\arguments{
\item{dirSim}{The simulation output folder}

\item{pathTrapGrid}{The path to the trap grd location}

\item{pathOut}{The path where the nc file should be saved}

\item{goodProj}{The projection of the target raster}

\item{shDoMix}{Should the mixing ratio be calculated or the First occurance?}

\item{shUseSum}{Should the procedure use the summation of the week or the snapshot?}
}
\value{
A raster of the selected analysis method. If the pathOut is specified,
Then the raster is outputed as an netCDF file with proper variable names and units
}
\description{
Either do the time of first Occurance, or the Mixing Ratio for either the
simulation or the difference between the simulation and the trap data.
Also can be used with either the snapshot view or the summation for the whole
week.
}

