% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scrubValidation.R
\name{scrubHaplo}
\alias{scrubHaplo}
\title{Make the haplotype data easier to look at}
\usage{
scrubHaplo(pathXlsx, pathTrapDict, pathCsvOut = "", maskBefore2011 = TRUE,
  maskYr = c(), maskNumMoth = 15)
}
\arguments{
\item{pathXlsx}{The path to the xlsx with the Haplotype data}

\item{pathTrapDict}{A csv with the translation of all the County names to the latLon location}

\item{pathCsvOut}{The csv location where the nice data should be outputed}

\item{maskBefore2011}{should values before 2011 be used?}

\item{maskYr}{An array of the years after 2010 that you want to mask}

\item{maskNumMoth}{The minimum number of moths that are required for a proper Haplotype experiment}
}
\value{
The final matrix and a csv file that was written if pathCsvOut was specified
}
\description{
Make the haplotype data easier to look at
}

